requirements.txt
# Core dependencies
fastmcp>=2.7.1
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OAuth and authentication
authlib>=1.3.0
cryptography>=41.0.0
aiohttp>=3.9.0
PyJWT>=2.8.0

# HTTP Server dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
sse-starlette>=2.0.0
python-jose>=3.3.0

# Async support
aiofiles>=23.0.0

# Timezone support
pytz>=2024.1

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
flake8>=6.1.0
mypy>=1.5.0
black>=23.9.0

# Type stubs
types-aiofiles>=23.0.0
types-pytz>=2024.1